feat: add content to first recipe page and add link to the commit convention used for odin-recipes in README.md
in lasagna.html, an image was added under the h1 heading, then a h2 heading and a short description of the dish was given, after that a h2 heading and an unordered list of ingredients was given and after that a hh2 heading and an ordered list of steps was given. Then, a link to the original website of the recipe was given. THen the link to the commit convention that this repository is following was added to the README.md file