feat: format recipe list of main page
the div named recipe-list that i made in my previous commits was styled in this commit using css. margin-left moved the div element to the center of the page while keeping the left alignment of the list. padding was given to create distance between the border and the list. width determined the width of the borders